emoney_wallet/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose
pydantic
